[BB] Updating the file to accept thresholds
This change accepts thresholds {[Warning] and [Critical] or none } and returns the exit status by comparing the count of "vhost, exchange, binding, queue, channel" with the thresholds which are passed as an input. Executed this change on command line and I could get valid results. Performed sanity checks with all possible approaches and could get expected output.