check_rabbitmq_objects: reduce amount of metadata requested
Parsing big JSON documents can take a long time, and we only want
the number of objects.  "name" is a valid attribute for all
endpoints except bindings, so we ask for "vhost" instead for that case.

On our OpenStack using RabbitMQ, CPU time for this check was reduced
from 52 seconds to 2 seconds.